Create mDNA_translate_orf.py
Chooses the longest open reading frame from mDNA and translate it to AA sequence.
Input parameters are: 
input_folder - the location of a folder where multiple fasta files with DNA sequences are
output_folder - the location (will create folder if not exists) of a folder where output will go. Files will be created with the same name as input.